refactor: shared component for FormView and QuickTimerView
