+++ Taskito/Timer/QuickTimerView.swift
@@ -17,8 +17,6 @@ struct QuickTimerView: View {
     @AppStorage("latestMinutesSelection") private var minutes = 25
     @AppStorage("latestSecondsSelection") private var seconds = 0
 
-    @FocusState private var focusedField: FocusedField?
-
     @State private var pendingTimerMinutes: Int?
     @State private var pendingTimerSeconds: Int?
 
@@ -67,69 +65,18 @@ struct QuickTimerView: View {
                 }
             } else {
                 // Timer setup view
-                VStack(spacing: 24) {
-                    // Timer input
-                    HStack(spacing: 12) {
-                        VStack(spacing: 8) {
-                            Text("Minutes")
-                                .font(.system(size: 12, weight: .semibold, design: .rounded))
-                                .foregroundStyle(.secondary)
-
-                            TextField("0", value: $minutes, formatter: timerPolicy.formatter)
-                                .textFieldStyle(.roundedBorder)
-                                .font(.system(size: 48, weight: .medium, design: .rounded))
-                                .multilineTextAlignment(.center)
-                                .frame(width: 100)
-                                .focused($focusedField, equals: .minutes)
-                        }
-
-                        Text(":")
-                            .font(.system(size: 48, weight: .thin, design: .rounded))
-                            .foregroundStyle(.secondary)
-                            .padding(.top, 20)
-
-                        VStack(spacing: 8) {
-                            Text("Seconds")
-                                .font(.system(size: 12, weight: .semibold, design: .rounded))
-                                .foregroundStyle(.secondary)
-
-                            TextField("0", value: $seconds, formatter: timerPolicy.formatter)
-                                .textFieldStyle(.roundedBorder)
-                                .font(.system(size: 48, weight: .medium, design: .rounded))
-                                .multilineTextAlignment(.center)
-                                .frame(width: 100)
-                                .focused($focusedField, equals: .seconds)
-                        }
-                    }
-
-                    // Start button
-                    Button(action: {
-                        startTimer(minutes: minutes, seconds: seconds)
-                    }) {
-                        Text("START")
-                            .font(.system(size: 14, weight: .semibold, design: .rounded))
-                            .tracking(0.5)
-                            .padding(.vertical, 12)
-                            .padding(.horizontal, 50)
-                            .background(
-                                Capsule(style: .continuous)
-                                    .fill(Color.accentColor)
-                            )
-                            .foregroundStyle(.white)
-                    }
-                    .keyboardShortcut(.defaultAction)
-                    .buttonStyle(.plain)
+                TimerFormView(
+                    minutes: $minutes,
+                    seconds: $seconds,
+                    timerPolicy: timerPolicy,
+                    variant: .medium
+                ) {
+                    startTimer(minutes: minutes, seconds: seconds)
                 }
             }
         }
         .padding(24)
         .frame(width: 300)
-        .onAppear {
-            // Auto-focus minutes field when view appears
-            DispatchQueue.main.async {
-                focusedField = .minutes
-            }
-        }
         .background(
             // ESC keyboard shortcut to close window
             Button("") {
@@ -141,33 +88,9 @@ struct QuickTimerView: View {
     }
 
     private func startTimer(minutes: Int, seconds: Int) {
-        // Check if a timer is already running
-        if countdownStore.timerState == .running || countdownStore.timerState == .paused {
-            // Store pending timer and show alert
-            pendingTimerMinutes = minutes
-            pendingTimerSeconds = seconds
-        } else {
-            // No conflict, start immediately
-            countdownStore.startNewTimer(minutes: minutes, seconds: seconds)
-            dismiss()
-        }
-    }
-
-    private func confirmStartTimer() {
-        guard let minutes = pendingTimerMinutes,
-              let seconds = pendingTimerSeconds else { return }
-
         countdownStore.startNewTimer(minutes: minutes, seconds: seconds)
-
-        // Clear pending state
-        pendingTimerMinutes = nil
-        pendingTimerSeconds = nil
         dismiss()
     }
-}
-
-private enum FocusedField: Hashable {
-    case minutes, seconds
 }
 
 #Preview {
